FEAT: add support for types in domains (#4)
* fix imports in models

* add domain type

* add vs code debugging

* vs code debugging ember command

* add type support in interpretation for plus and minus nodes

* add type support interpretation for all nodes

* promts before delete

* new dump

+++ frontend/app/controllers/frame-base/domains.ts
@@ -6,12 +6,16 @@ import { Domain, DomainValue, FrameBase } from "knowledge-shell/models";
 export default class FrameBaseDomains extends Controller {
   @tracked search = "";
 
-  @computed.oneWay("model")
+  @computed.oneWay("model") 
   frameBase!: FrameBase;
 
+  @computed("model.domains.[]")
   get orderedDomains(): Domain[] {
-    return this.model.domains.sortBy("name");
-  }
+    const frameDomain = this.frameBase.frameDomain;
+    const domains = this.frameBase.domains.filter((domain: Domain) => !domain.isReadOnly);
+    const sortedDomains = domains.sortBy("name");
+    return [frameDomain].concat(sortedDomains);
+  };
 
   @action
   addDomain() {
@@ -39,12 +43,15 @@ export default class FrameBaseDomains extends Controller {
 
   @action
   deleteDomain(domain: Domain): void {
-    domain.domainValues.forEach((domainValue: DomainValue) => {
-      domainValue.destroyRecord();
-    });
-    domain.destroyRecord();
+    const shouldBeDeleted = window.confirm(`Are you sure you want to delete ${domain.name}?`);
+    if (shouldBeDeleted) {
+      domain.domainValues.forEach((domainValue: DomainValue) => {
+        domainValue.destroyRecord();
+      });
+      this.frameBase.domains.removeObject(domain);
+      domain.destroyRecord();
+    }
   }
-
 }
 
 declare module "@ember/controller" {
